refactoring and cleanup
- moved folder structure for clarity
- add telegram webapp typing
- disable postman ci integration test
- fix some eslint errors
