+++ .gitignore
@@ -73,3 +73,7 @@ local-data
 
 # testing data
 testing-data/**
+
+# Others
+.idea
+.DS_Store
