+++ checkers-app/src/components/common/BackButton.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@material-tailwind/react";
 import { ArrowLeftIcon } from "@heroicons/react/24/solid";
 import { useNavigate } from "react-router-dom";
-import { useUser } from "../UserContext";
+import { useUser } from "../../providers/UserContext";
 
 export function BackButton() {
   const { phoneNo } = useUser();
