+++ checkers-app/src/components/common/Header.tsx
