+++ checkers-app/src/components/common/Layout.tsx
@@ -1,20 +1,20 @@
-import { ReactNode} from 'react';
+import { ReactNode } from 'react';
 import Header from "./Header";
 import NavbarDefault from "./BotNavBar";
-import { useUser } from '../UserContext';
+import { useUser } from '../../providers/UserContext';
 
 interface LayoutProps {
     pageHeader: string,
     children: ReactNode,
 }
 
-export default function Layout({pageHeader, children }: LayoutProps) {
-    const {unassessed, unchecked} = useUser();
-    
+export default function Layout({ pageHeader, children }: LayoutProps) {
+    const { unassessed, unchecked } = useUser();
+
     return (
         <div>
             <div className='layout-padding'>
-                <Header name={pageHeader}/>
+                <Header name={pageHeader} />
                 {/* <PageHeader>{pageHeader}</PageHeader> */}
             </div>
             <div className="pb-16">
