+++ checkers-app/src/components/common/MessageCard.tsx
