+++ checkers-app/src/components/common/PageHeader.tsx
