+++ checkers-app/src/pages/DashboardPage.tsx
@@ -1,9 +1,9 @@
 import Dashboard from "../components/dashboard";
-import Layout from "../shared/Layout";
+import Layout from "../components/common/Layout";
 
-export default function DashboardPage (){
+export default function DashboardPage() {
     return (
-    <Layout pageHeader="DASHBOARD">
-        <Dashboard/> 
-    </Layout>)
+        <Layout pageHeader="DASHBOARD">
+            <Dashboard />
+        </Layout>)
 }
