+++ checkers-app/src/pages/MyVotesPage.tsx
@@ -1,5 +1,5 @@
 import MyVotes from "../components/myvotes";
-import Layout from "../shared/Layout";
+import Layout from "../components/common/Layout";
 
 export default function MyVotesPage() {
     return (
